move and rename method isWithinOneHourBeforeOpeningTime, rename exception, add ClientFinalPlaceController
