--- src/main/java/dk/cngroup/lentils/exception/NotOneHourBeforeOpeningTimeException.java
@@ -1,7 +1,0 @@
-package dk.cngroup.lentils.exception;
-
-public class NotOneHourBeforeOpeningTimeException extends RuntimeException {
-    public NotOneHourBeforeOpeningTimeException() {
-        super("Wait for the right time...");
-    }
-}
