+++ src/main/java/dk/cngroup/lentils/exception/NotYetTimeToShowOpeningTimeException.java
@@ -1,0 +1,7 @@
+package dk.cngroup.lentils.exception;
+
+public class NotYetTimeToShowOpeningTimeException extends RuntimeException {
+    public NotYetTimeToShowOpeningTimeException() {
+        super("Wait for the right time...");
+    }
+}
